project-1_[telegram-bot-uni]/requirements.txt
aiogram==2.11.2
aiohttp>=3.7.3
aioredis==1.3.1
python-dotenv~=0.15.0
psycopg2~=2.8.6
peewee~=3.14.0
wasabi~=0.8.2